Fix scheduler stack :recycle:
The pile goes down I forgot that

+++ Kernel/src/scheduling/scheduler/scheduler.cpp
@@ -2,9 +2,8 @@
 
 TaskManager globalTaskManager;
 
-void TaskManager::Scheduler(struct InterruptStack* Registers, uint8_t CoreID){    
-    if(CoreInUserSpace[CoreID]){ 
-
+void TaskManager::Scheduler(InterruptStack* Registers, uint8_t CoreID){   
+    if(CoreInUserSpace[CoreID]){   
         TaskNode* node = NodeExecutePerCore[CoreID];
 
         node->Content.Regs.rax = Registers->rax;
@@ -27,42 +26,36 @@ void TaskManager::Scheduler(struct InterruptStack* Registers, uint8_t CoreID){
         node->Content.Regs.rsp = Registers->rsp;
         node->Content.Regs.ss = Registers->ss;
 
-        /*printf("%k %u %k", 0xffffff00, MainNodeScheduler->Content.ID, 0xffffffff);
-        if(MainNodeScheduler->Next == NULL){
-            MainNodeScheduler = FirstNode;
-        }else{
-            MainNodeScheduler = MainNodeScheduler->Next;
-        }
-        
-        NodeExecutePerCore[CoreID] = MainNodeScheduler;
+        MainNodeScheduler = MainNodeScheduler->Next;   
+        node = MainNodeScheduler;
+        NodeExecutePerCore[CoreID] = node;
 
-
-        Registers->rax = MainNodeScheduler->Content.Regs.rax;
-        Registers->rcx = MainNodeScheduler->Content.Regs.rcx;
-        Registers->rdx = MainNodeScheduler->Content.Regs.rdx;
-        Registers->rsi = MainNodeScheduler->Content.Regs.rsi;
-        Registers->rdi = MainNodeScheduler->Content.Regs.rdi;
-        Registers->rbp = MainNodeScheduler->Content.Regs.rbp;
-        Registers->r8 = MainNodeScheduler->Content.Regs.r8;
-        Registers->r9 = MainNodeScheduler->Content.Regs.r9;
-        Registers->r10 = MainNodeScheduler->Content.Regs.r10;
-        Registers->r11 = MainNodeScheduler->Content.Regs.r11;
-        Registers->r12 = MainNodeScheduler->Content.Regs.r12;
-        Registers->r13 = MainNodeScheduler->Content.Regs.r13;
-        Registers->r14 = MainNodeScheduler->Content.Regs.r14;
-        Registers->r15 = MainNodeScheduler->Content.Regs.r15;
-        Registers->rip = MainNodeScheduler->Content.Regs.rip;
-        Registers->cs = MainNodeScheduler->Content.Regs.cs;
-        Registers->rflags = MainNodeScheduler->Content.Regs.rflags;
-        Registers->rsp = MainNodeScheduler->Content.Regs.rsp;
-        Registers->ss = MainNodeScheduler->Content.Regs.ss;*/
+        Registers->rax = node->Content.Regs.rax;
+        Registers->rcx = node->Content.Regs.rcx;
+        Registers->rdx = node->Content.Regs.rdx;
+        Registers->rsi = node->Content.Regs.rsi;
+        Registers->rdi = node->Content.Regs.rdi;
+        Registers->rbp = node->Content.Regs.rbp;
+        Registers->r8 = node->Content.Regs.r8;
+        Registers->r9 = node->Content.Regs.r9;
+        Registers->r10 = node->Content.Regs.r10;
+        Registers->r11 = node->Content.Regs.r11;
+        Registers->r12 = node->Content.Regs.r12;
+        Registers->r13 = node->Content.Regs.r13;
+        Registers->r14 = node->Content.Regs.r14;
+        Registers->r15 = node->Content.Regs.r15;
+        Registers->rip = node->Content.Regs.rip;
+        Registers->cs = node->Content.Regs.cs;
+        Registers->rflags = node->Content.Regs.rflags;
+        Registers->rsp = node->Content.Regs.rsp;
+        Registers->ss = node->Content.Regs.ss;
     }
 }
 
-TaskNode* TaskManager::AddTask(void* EntryPoint, size_t Size, bool IsIddle){ 
+TaskNode* TaskManager::AddTask(void* EntryPoint, size_t Size, bool IsIddle, bool IsLinked){ 
     TaskNode* node = (TaskNode*)malloc(sizeof(TaskNode));
     
-    uint64_t StackSize = sizeof(ContextStack);
+    uint64_t StackSize = 0x4000;
     node->Content.Stack = malloc(StackSize);
     for(int i = 0; i < (StackSize / 0x1000) + 1; i++){
         globalPageTableManager.MapUserspaceMemory((void*)((uint64_t)node->Content.Stack + i * 0x1000));
@@ -76,23 +69,25 @@ TaskNode* TaskManager::AddTask(void* EntryPoint, size_t Size, bool IsIddle){
     node->Content.Regs.rip = EntryPoint; 
     node->Content.Regs.cs = (void*)GDTInfoSelectors.UCode; //user code selector
     node->Content.Regs.ss = (void*)GDTInfoSelectors.UData; //user data selector
-    node->Content.Regs.rsp = node->Content.Stack;
+    node->Content.Regs.rsp = (void*)((uint64_t)node->Content.Stack + StackSize); //because the pile goes down I had not seen it ;(
     node->Content.Regs.rflags = (void*)0x202; //interrupts & syscall
     node->Content.IsIddle = IsIddle;
-    
     
     node->Content.ID = IDTask; //min of ID is 0
     IDTask++;
     NumTaskTotal++;
     //link
-    node = NewNode(node);
-    if(IsIddle){  
-        IdleNode[IddleTaskNumber] = node;
+    if(IsLinked){
+        node = NewNode(node);
+        if(IsIddle){  
+            IdleNode[IddleTaskNumber] = node;
 
-        IddleTaskNumber++;
-    }else if(IddleTaskNumber != 0){
-        DeleteTask(IdleNode[IddleTaskNumber - 1]); /* because we add 1 in this function */
+            IddleTaskNumber++;
+        }else if(IddleTaskNumber != 0){
+            DeleteTask(IdleNode[IddleTaskNumber - 1]); /* because we add 1 in this function */
+        }        
     }
+
 
     return node;
 }
@@ -107,13 +102,15 @@ TaskNode* TaskManager::NewNode(TaskNode* node){
         LastNode->Next = node;
     }
 
+    node->Next = FirstNode;
     MainNode = node;
     LastNode = node;
+    
     return node;
 }
 
-TaskNode* TaskManager::CreatDefaultTask(){
-    TaskNode* node = globalTaskManager.AddTask((void*)IdleTask, 0x1000, true);
+TaskNode* TaskManager::CreatDefaultTask(bool IsLinked){
+    TaskNode* node = globalTaskManager.AddTask((void*)IdleTask, 0x1000, true, IsLinked);
 }
 
 void TaskManager::DeleteTask(TaskNode* node){
@@ -124,11 +121,32 @@ void TaskManager::DeleteTask(TaskNode* node){
     TaskNode* next = node->Next;
     TaskNode* last = node->Last;
 
-    if(node == MainNode || node == LastNode || node == FirstNode || node == MainNodeScheduler){
+    if(node == MainNode){
         if(next != NULL){
             MainNode = next;
         }else{
             MainNode = last;
+        }
+    }
+    if(node == LastNode){
+        if(next != NULL){
+            LastNode = next;
+        }else{
+            LastNode = last;
+        }
+    }
+    if(node == FirstNode){
+        if(last != NULL){
+            FirstNode = last;
+        }else{
+            FirstNode = next;
+        }
+    }
+    if(node == MainNodeScheduler){
+        if(next != NULL){
+            MainNodeScheduler = next;
+        }else{
+            MainNodeScheduler = last;
         }
     }
 
@@ -140,28 +158,29 @@ void TaskManager::DeleteTask(TaskNode* node){
 
     if(NumTaskTotal <= APIC::ProcessorCount){
         IdleNode[IddleTaskNumber] = NULL;
-        CreatDefaultTask();
+        CreatDefaultTask(true);
     }
 }
 
 void TaskManager::InitScheduler(uint8_t NumberOfCores){
     for(int i = 0; i < NumberOfCores; i++){
-        CreatDefaultTask();
-    }   
+        CreatDefaultTask(true);
+    } 
 
     TaskManagerInit = true;
 }
 
 void TaskManager::EnabledScheduler(uint8_t CoreID){ 
     if(TaskManagerInit){
-        TaskNode* node = MainNodeScheduler;
-        MainNodeScheduler = MainNodeScheduler->Next;
+        CoreInUserSpace[CoreID] = true;
 
         EnableSystemCall(); 
         
-        NodeExecutePerCore[CoreID] = node;
+        TaskNode* node = CreatDefaultTask(false);
 
-        JumpIntoUserspace(node->Content.EntryPoint, node->Content.Stack, node->Content.Regs.cs, node->Content.Regs.ss, CoreID); 
+        NodeExecutePerCore[CoreID] = node;
+        
+        JumpIntoUserspace(node->Content.EntryPoint, node->Content.Stack, node->Content.Regs.cs, node->Content.Regs.ss, (uint64_t)CoreID); 
     }
 
 }
